change names of navG for cause i was confused but made no difference, just need to comment out lines on navReo and navEng to make it work
