+++ client/components/App.jsx
@@ -4,7 +4,7 @@ import { HashRouter as Router, Route } from "react-router-dom";
 import { isAuthenticated } from "../utils/auth"
 
 import Landing from "./LandingG/Landing";
-import Nav from "./NavG/Nav";
+import Toggle from "./ToggleG/Toggle";
 import Explore from "./ExploreG/Explore";
 
 import AddRahuiMap from "./NewMap";
@@ -29,6 +29,7 @@ import FormAndMap from "./FormAndMap";
 import Register from "./Register";
 import Tautoko from "./Tautoko";
 import Project from "./ProjectG/Project";
+import Nav from "./Nav"
 
 // import Register from "./Register"
 
@@ -36,7 +37,9 @@ export function App({ auth }) {
   return (
     <div>
       <Router>
+
         <Nav />
+        <Toggle />
         <Route exact path="/" component={Landing} />
 
         <Route path="/about" component={About} />
